Bug fix:  Event list not rendering after number of events input
When the user opened the app and selected a city or show all cities, the app is displaying the default number of events. When changing the number of events did not render anymore (only after selecting the city or all cities again) Expected behavior: Rendering the number of events  everytime there was an input.